requirements.txt
faiss-cpu
sentence-transformers
numpy
streamlit
PyPDF2
python-docx
langchain-google-genai
langgraph
regex